+++ example/lib/retry_example.dart
@@ -27,6 +27,17 @@ void main() {
   // of LoadingStateHandlerWidget in the app
   LoadingStateHandlerWidget.setDefaults(
     defaultRetryCooldown: const Duration(seconds: 5),
+    defaultErrorBuilder:
+        (context, errorMessage, retryWidget, retryCooldown, onRetry) => Center(
+      child: Column(
+        spacing: 20,
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(errorMessage ?? 'An error occurred'),
+          retryWidget,
+        ],
+      ),
+    ),
   );
 
   runApp(
@@ -135,23 +146,23 @@ class _RetryExampleState extends State<RetryExample> {
     await Future.delayed(const Duration(seconds: 2));
 
     // Simulate random failure scenario
-    if (DateTime.now().second % 2 == 0) {
-      setState(() {
-        _isLoading = false;
-        _hasError = true;
-        _errorMessage = 'Failed to fetch data. Please try again.';
-      });
-      return;
-    }
+    // if (DateTime.now().second % 2 == 0) {
+    setState(() {
+      _isLoading = false;
+      _hasError = true;
+      _errorMessage = 'Failed to fetch data. Please try again.';
+    });
+    // return;
+    // }
 
-    // Simulate empty response scenario
-    if (DateTime.now().second % 3 == 0) {
-      setState(() {
-        _isLoading = false;
-        _isEmpty = true;
-      });
-      return;
-    }
+    // // Simulate empty response scenario
+    // if (DateTime.now().second % 3 == 0) {
+    //   setState(() {
+    //     _isLoading = false;
+    //     _isEmpty = true;
+    //   });
+    //   return;
+    // }
 
     // Success case - data loaded successfully
     setState(() {